requirements.txt
requests>=2.32.0
python-dotenv>=1.0.0
psycopg[binary]>=3.2.0
